requirements.txt
gevent_socketio==0.3.6
numpy==1.20.3
opencv_python==4.5.5.64
pynput==1.7.6
